feat: Added auth layer for organisation

+++ api/domain/organisation.py
@@ -1,6 +1,11 @@
 from sqlalchemy import Engine
 from sqlmodel import Session
 
+from domain.db_utils import handle_db_operation
+from domain.exceptions import (
+    OrganisationAccessDeniedError,
+    OrganisationNotFoundError,
+)
 from models.base import Organisation
 from models.schemas import OrganisationSchema
 
@@ -9,28 +14,43 @@ class OrganisationService:
     def __init__(self, client: Engine):
         self.client = client
 
-    def create_organisation(self, organistaion: Organisation) -> OrganisationSchema:
+    @handle_db_operation
+    def create_organisation(
+        self, owner_id: str, organistaion: Organisation
+    ) -> OrganisationSchema:
         with Session(self.client) as session:
-            organisation_schema = OrganisationSchema.model_validate(organistaion)
+            # Check if organization with same name exists for this owneer
+            print(organistaion)
+            to_create = organistaion.model_dump()
+            to_create.update({"owner_id": owner_id})
+            organisation_schema = OrganisationSchema.model_validate(to_create)
             session.add(organisation_schema)
             session.commit()
             session.refresh(organisation_schema)
             return organisation_schema
 
-    def get_organisation(self, organisation_id: str) -> OrganisationSchema:
-        with Session(self.client) as session:
-            organisation = session.get(OrganisationSchema, organisation_id)
-            if not organisation:
-                return None
-            return organisation
-
-    def update_organisation(
-        self, organisation_id: str, organisation_data: Organisation
+    @handle_db_operation
+    def get_organisation(
+        self, owner_id: str, organisation_id: str
     ) -> OrganisationSchema:
         with Session(self.client) as session:
             organisation = session.get(OrganisationSchema, organisation_id)
             if not organisation:
-                return None
+                raise OrganisationNotFoundError(organisation_id)
+            if organisation.owner_id != owner_id:
+                raise OrganisationAccessDeniedError(organisation_id, owner_id)
+            return organisation
+
+    @handle_db_operation
+    def update_organisation(
+        self, owner_id: str, organisation_id: str, organisation_data: Organisation
+    ) -> OrganisationSchema:
+        with Session(self.client) as session:
+            organisation = session.get(OrganisationSchema, organisation_id)
+            if not organisation:
+                raise OrganisationNotFoundError(organisation_id)
+            if organisation.owner_id != owner_id:
+                raise OrganisationAccessDeniedError(organisation_id, owner_id)
             for key, value in organisation_data.model_dump().items():
                 setattr(organisation, key, value)
             session.add(organisation)
@@ -38,11 +58,14 @@ class OrganisationService:
             session.refresh(organisation)
             return organisation
 
-    def delete_organisation(self, organisation_id: str) -> bool:
+    @handle_db_operation
+    def delete_organisation(self, owner_id: str, organisation_id: str) -> bool:
         with Session(self.client) as session:
             organisation = session.get(OrganisationSchema, organisation_id)
             if not organisation:
-                return False
+                raise OrganisationNotFoundError(organisation_id)
+            if organisation.owner_id != owner_id:
+                raise OrganisationAccessDeniedError(organisation_id, owner_id)
             session.delete(organisation)
             session.commit()
             return True
